Add files via upload
On a sultry Friday morning in the month of August Mr Suresh Babu, head of the Steel melting shop in Tata Steel, was glancing through the daily report of his department with a brow of concern on his forehead.
In the last couple of days 6 heats were downgraded due to high Sulphur content. He was now worried about meeting the delivery schedule of a premium order from one of their highly esteemed customers.
He was also concerned that, Sales team will have difficulty in finding an order for the downgraded heats - which will eventually be sold at a lower price - directly affecting the bottom line.
Sulphur content in steel is currently being predicted by a model and the reason for Mr Babu’s immediate concern was that despite his team’s best efforts his current model only has a 65% accuracy of predicting the Sulphur % in the given specification for few grades. Given his habit of staying abreast of the advances in the field of Predictive analytics, Mr Babu was of the firm belief that an analytics led intervention may act as a panacea for the problem he was currently encountering.
Mr Babu is looking for a team to help him better predict the Output Sulphur % of his Desulphurization process?
In order to aid the team Mr Suresh Babu has listed down a brief of the process and the objectives he wanted the
Analytical solution to address
1.0 Process Description
To achieve low Sulphur content steel is Desulphurized in a desulphurization(DS) unit.
The Desulphurization unit receives hot metal with high Sulphur level (0.02-0.1 %) from the blast furnaces. The
hot metal is then treated to reduce the Sulphur level to 0.002-0.015 % before sending it to LD converter for further
processing.
Desulphurization is carried out by the injection process, in which Calcium carbide and Magnesium powder
(DS compound) are mixed in a fixed ratio and injected into liquid hot metal through a refractory top lance with an
inert carrier gas. The injection time varies from 5 to 20 minutes. After the completion of desulphurization, the slag
formed on the top is highly rich in Sulphur and it is skimmed out completely to prevent any Sulphur reversal during
primary steel making.
After the completion of desulphurization, the slag formed on the top is highly rich in Sulphur and it is
skimmed out completely to prevent any Sulphur reversal during primary steel making.